Moved database row expiration mechanism to a trait to avoid code duplication.
