+++ module/VuFind/src/VuFind/Db/Table/ExternalSession.php
@@ -28,7 +28,6 @@
  * @link     https://vufind.org Main Page
  */
 namespace VuFind\Db\Table;
-use Zend\Db\Sql\Expression;
 
 /**
  * Table Definition for external_session
@@ -42,6 +41,8 @@ use Zend\Db\Sql\Expression;
  */
 class ExternalSession extends Gateway
 {
+    use ExpirationTrait;
+
     /**
      * Constructor
      */
@@ -93,57 +94,25 @@ class ExternalSession extends Gateway
     }
 
     /**
-     * Delete expired sessions. Allows setting of 'from' and 'to' ID's so that rows
-     * can be deleted in small batches.
+     * Update the select statement to find records to delete.
      *
-     * @param int $daysOld Age in days of an "expired" session.
-     * @param int $idFrom  Lowest id of rows to delete.
-     * @param int $idTo    Highest id of rows to delete.
+     * @param Select $select  Select clause
+     * @param int    $daysOld Age in days of an "expired" record.
+     * @param int    $idFrom  Lowest id of rows to delete.
+     * @param int    $idTo    Highest id of rows to delete.
      *
-     * @return int Number of rows deleted
+     * @return void
      */
-    public function deleteExpired($daysOld = 2, $idFrom = null, $idTo = null)
-    {
+    protected function expirationCallback($select, $daysOld, $idFrom = null,
+        $idTo = null
+    ) {
         $expireDate = date('Y-m-d', time() - $daysOld * 24 * 60 * 60);
-        $callback = function ($select) use ($expireDate, $idFrom, $idTo) {
-            $where = $select->where->lessThan('created', $expireDate);
-            if (null !== $idFrom) {
-                $where->and->greaterThanOrEqualTo('id', $idFrom);
-            }
-            if (null !== $idTo) {
-                $where->and->lessThanOrEqualTo('id', $idTo);
-            }
-        };
-        return $this->delete($callback);
-    }
-
-    /**
-     * Get the lowest id and highest id for expired sessions.
-     *
-     * @param int $daysOld Age in days of an "expired" session.
-     *
-     * @return array|bool Array of lowest id and highest id or false if no expired
-     * records found
-     */
-    public function getExpiredIdRange($daysOld = 2)
-    {
-        $expireDate = date('Y-m-d', time() - $daysOld * 24 * 60 * 60);
-        $callback = function ($select) use ($expireDate) {
-            $select->where->lessThan('created', $expireDate);
-        };
-        $select = $this->getSql()->select();
-        $select->columns(
-            [
-                'id' => new Expression('1'), // required for TableGateway
-                'minId' => new Expression('MIN(id)'),
-                'maxId' => new Expression('MAX(id)'),
-            ]
-        );
-        $select->where($callback);
-        $result = $this->selectWith($select)->current();
-        if (null === $result->minId) {
-            return false;
+        $where = $select->where->lessThan('created', $expireDate);
+        if (null !== $idFrom) {
+            $where->and->greaterThanOrEqualTo('id', $idFrom);
         }
-        return [$result->minId, $result->maxId];
+        if (null !== $idTo) {
+            $where->and->lessThanOrEqualTo('id', $idTo);
+        }
     }
 }
